Fix: Ajout d'une bean pour les mappers et ajout de l'auto-increment

+++ officeTourismeAPI/src/main/java/com/officetourisme/dtos/CommentaireDto.java
@@ -11,11 +11,11 @@ import java.sql.Timestamp;
 @Setter
 @Getter
 public class CommentaireDto {
-    private Integer id;
+    private Long id;
     private int type;
     @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss")
     private Timestamp date;
     private String contenu;
-    private Integer compteId;
-    private Integer sortieId;
+    private Long compteId;
+    private Long sortieId;
 }
