+++ officeTourismeAPI/src/main/java/com/officetourisme/repositories/CommentaireRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentaireRepository extends JpaRepository<Commentaire, Integer> {
+public interface CommentaireRepository extends JpaRepository<Commentaire, Long> {
 }
