+++ officeTourismeAPI/src/main/java/com/officetourisme/repositories/PanierRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PanierRepository extends JpaRepository<Panier, Integer> {
+public interface PanierRepository extends JpaRepository<Panier, Long> {
 }
