+++ officeTourismeAPI/src/main/java/com/officetourisme/repositories/SortieRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SortieRepository extends JpaRepository<Sortie, Integer> {
+public interface SortieRepository extends JpaRepository<Sortie, Long> {
 }
