+++ officeTourismeAPI/src/main/java/com/officetourisme/services/PanierService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface PanierService {
     PanierDto savePanier(PanierDto panierDto);
-    PanierDto getPanierById(Integer panierId);
-    boolean deletePanier(Integer panierId);
+    PanierDto getPanierById(Long panierId);
+    boolean deletePanier(Long panierId);
     List<PanierDto> getAllPaniers();
 }
